--- src/service/projectManagement/task/updateTask/updateTask.route.ts
@@ -1,15 +1,0 @@
-import { Router } from "express";
-import * as updateTaskController from "./updateTask.controller";
-import * as IPSpamChecker from "../../../middleware/api.limiter";
-import * as APIValidator from "../../../middleware/api.validator";
-
-const router = Router();
-
-router.put(
-  "/updateTask",
-  IPSpamChecker.checkIpSpamServer("/projectManagement"), 
-  APIValidator.taskValidator,
-  updateTaskController.updateTaskController,
-);
-
-export default router;
